requirements.txt
streamlit
torch
torchvision
numpy
pillow
opencv-python
plotly
streamlit-option-menu
